Move scheduler into scheduler package
