+++ src/main/kotlin/ar/pelotude/ohhsugoi/db/scheduler/DiscordHookMessage.kt
@@ -1,0 +1,6 @@
+package ar.pelotude.ohhsugoi.db.scheduler
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class DiscordHookMessage(val username: String="Sheska", val content: String)
